# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Poket Project API",
        "description": "API for managing user sets, see all cards and all sets of pokemon.",
        "contact": {
            "email": "hello@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user-sets": {
            "get": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Get all user sets",
                "operationId": "6518e4d95be9815a17cc467e792f3107",
                "responses": {
                    "200": {
                        "description": "List of user sets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserSet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Create a new user set",
                "operationId": "72761e3e9bb3a698c0b68ee134f44fe0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User set created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSet"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-sets/{userSet}": {
            "get": {
                "tags": [
                    "UserSets"
                ],
                "summary": "See a user set and cards of the set",
                "operationId": "d8ef00671b25db2c742f3c4ef1667215",
                "parameters": [
                    {
                        "name": "userSet",
                        "in": "path",
                        "description": "User set identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User set details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User set not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Update a user set",
                "operationId": "c89e50a5a7cecfe90369b7e2c3af96f8",
                "parameters": [
                    {
                        "name": "userSet",
                        "in": "path",
                        "description": "User set identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated user set",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSet"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Delete a user set",
                "operationId": "44e12b4c2770791a3ecee5507cbab3c4",
                "parameters": [
                    {
                        "name": "userSet",
                        "in": "path",
                        "description": "User set identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User set deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Set eliminado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-sets/{userSet}/add-card/{card}": {
            "post": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Add a card to a user set",
                "operationId": "0527af3e713c0b324a017d4139c8f8c6",
                "parameters": [
                    {
                        "name": "userSet",
                        "in": "path",
                        "description": "User set identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "card",
                        "in": "path",
                        "description": "Card identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carta añadida correctamente al set"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-sets/{userSet}/remove-card/{card}": {
            "delete": {
                "tags": [
                    "UserSets"
                ],
                "summary": "Remove a card from a user set",
                "operationId": "5890e257190a8c47022c553a54a5846e",
                "parameters": [
                    {
                        "name": "userSet",
                        "in": "path",
                        "description": "User set identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "card",
                        "in": "path",
                        "description": "Card identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carta eliminada correctamente del set"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserSet": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "card_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cards": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Card"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserSetRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "El nombre del set",
                        "type": "string",
                        "example": "Mi Set de Pokémon"
                    },
                    "description": {
                        "description": "Descripción del set",
                        "type": "string",
                        "example": "Un set especial con cartas raras."
                    },
                    "image": {
                        "description": "Imagen del set",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    }
                },
                "type": "object"
            }
        },
...
    },
...
}